feat: Add Tool Wizard for guided tool specification creation and editing
- Implemented ToolWizard screen with functionality to create and edit tool specifications.
- Added parameter management (add, edit, delete) within the wizard.
- Integrated JSON generation and file saving capabilities.
- Enhanced UI with appropriate styling and layout.

feat: Introduce Tool Manager for handling tool specifications

- Created ToolManager class to manage loading, saving, and deleting tool specifications.
- Added methods for formatting tool content with syntax highlighting.

style: Add CSS styles for TUI components

- Defined styles for main layout, editor modals, buttons, and form elements.
- Improved visual consistency across various UI components.

test: Implement unit tests for TUI components

- Added tests for ErrorDialog, SaveFileDialog, PropertyEditor, and ToolEditor.
- Ensured coverage for UI interactions and validation logic.

chore: Update main application entry point

- Configured logging setup before application starts.
- Handled exceptions gracefully with logging for better debugging.

chore: Update requirements and scripts

- Added dependencies for textual and rich libraries.
- Created scripts for running tests and managing the application.

docs: Add initial documentation and comments

- Included docstrings for new classes and methods.
- Updated README with usage instructions and examples.

refactor: Clean up and organize project structure

- Moved utility functions into a dedicated utils module.
- Removed obsolete files and directories from the web_app.
