+++ cli/internal/tool/core/property.py
