+++ cli/internal/tool/test_tool_initialization.py
