+++ cli/internal/tool/tests/test_tool_initialization.py
