+++ cli/internal/tui/components/status.py
