+++ cli/internal/tui/dialogs/error.py
