+++ cli/internal/tui/dialogs/file_simplified.py
