+++ cli/internal/tui/dialogs/test_error_dialog.py
