+++ cli/internal/tui/screens/property_editor.py
