+++ cli/internal/tui/screens/tool_editor.py
