+++ cli/internal/tui/tests/test_dialogs.py
