+++ cli/internal/tui/tests/test_property_editor.py
