+++ cli/internal/tui/tools/tool_manager.py
