+++ cli/screenshots/add_tool_parameter.jpg
