+++ cli/screenshots/edit_tool.jpg
