+++ cli/screenshots/edit_tool_property.jpg
