+++ cli/screenshots/main_screen.jpg
