+++ cli/screenshots/new_tool.jpg
