+++ cli/scripts/test.sh
