+++ web_app/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "web_app",
+  "name": "hero-forge",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.0.1",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite --host",
     "build": "tsc -b && vite build",
     "lint": "eslint .",
     "preview": "vite preview"
